use own string implementation, add more movement, add delete-line
The string implementation is a vector of strings that keeps the total
length of its data, plus the newline count. There's no good way to
retrieve a line (or its starting offset, or its length, ...) given a
character offset. Also, general insert and delete operations will be
annoying to implement. In other words, this implementation totally sucks
and exists only as a placeholder that should be replaced by a rope
implementation.